+++ packages/frontend/src/components/rosette/index.ts
@@ -1,3 +1,2 @@
 export * from './Rosette'
 export * from './TooltipPopup'
-export * from './types'
