+++ packages/frontend/src/pages/scaling-projects/props/getProjectHeader.ts
@@ -2,11 +2,11 @@ import { Layer2, ProjectLinks } from '@l2beat/config'
 import { ActivityApiResponse, TvlApiResponse } from '@l2beat/shared'
 
 import { Config } from '../../../build/config'
-import { RiskSentiments } from '../../../components/rosette'
 import { formatLargeNumber } from '../../../utils'
 import { getTpsDaily } from '../../../utils/activity/getTpsDaily'
 import { getTpsWeeklyChange } from '../../../utils/activity/getTpsWeeklyChange'
 import { getTransactionCount } from '../../../utils/activity/getTransactionCount'
+import { getRiskSentiments } from '../../../utils/risks/values'
 import { getTvlBreakdown } from '../../../utils/tvl/getTVLBreakdown'
 import { getTvlWithChange } from '../../../utils/tvl/getTvlWitchChange'
 import { formatUSD } from '../../../utils/utils'
@@ -52,7 +52,8 @@ export function getProjectHeader(
     tvlBreakdown,
     links: getLinks(project.display.links),
     ratingEntry: config.features.rating && project.rating,
-    risks: getRiskSentimentsFromRiskView(project.riskView),
+    // TODO: will need to be riskValues when rosette has hover
+    risks: getRiskSentiments(project.riskView),
     isArchived: project.isArchived,
     isUpcoming: project.isUpcoming,
   }
@@ -70,18 +71,6 @@ function getTitleLength(name: string): 'long' | 'very-long' | undefined {
     case 'Polygon Hermez':
     case 'Metis Andromeda':
       return 'very-long'
-  }
-}
-
-function getRiskSentimentsFromRiskView(
-  riskView: Layer2['riskView'],
-): RiskSentiments {
-  return {
-    sequencerFailure: riskView.sequencerFailure.sentiment,
-    stateValidation: riskView.stateValidation.sentiment,
-    dataAvailability: riskView.dataAvailability.sentiment,
-    upgradeability: riskView.upgradeability.sentiment,
-    validatorFailure: riskView.validatorFailure.sentiment,
   }
 }
