+++ packages/frontend/src/utils/risks/types.ts
@@ -7,11 +7,6 @@ type Risk =
   | 'validatorFailure'
 export type RiskSentiments = Partial<Record<Risk, Sentiment>>
 
-export interface RosetteProps {
-  risks: RiskSentiments
-  className?: string
-}
-
 export interface RiskValue {
   value: string
   sentiment?: Sentiment
