+++ packages/frontend/src/utils/risks/values.ts
@@ -1,0 +1,27 @@
+import { Layer2, Layer2RiskView } from '@l2beat/config'
+
+import { RiskSentiments, RiskValues } from './types'
+
+export function getRiskValuesFromRiskView(
+  riskView: Layer2RiskView,
+): RiskValues {
+  return {
+    dataAvailability: riskView.dataAvailability,
+    sequencerFailure: riskView.sequencerFailure,
+    stateValidation: riskView.stateValidation,
+    upgradeability: riskView.upgradeability,
+    validatorFailure: riskView.validatorFailure,
+  }
+}
+
+export function getRiskSentiments(
+  riskView: Layer2['riskView'] | RiskValues,
+): RiskSentiments {
+  return {
+    sequencerFailure: riskView.sequencerFailure.sentiment,
+    stateValidation: riskView.stateValidation.sentiment,
+    dataAvailability: riskView.dataAvailability.sentiment,
+    upgradeability: riskView.upgradeability.sentiment,
+    validatorFailure: riskView.validatorFailure.sentiment,
+  }
+}
